aexad/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
cycler==0.11.0
decorator==4.4.2
flatbuffers==1.12
gast==0.3.3
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==6.3.0
joblib==1.2.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.5.3
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.4.2
networkx==2.5.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==23.1
Pillow==8.3.2
protobuf==3.19.6
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.1
scikit-learn==1.5.2
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.2
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing-extensions==4.12.2
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
